pyproject.toml
[project]
name = "yt-thumbnail-abtest"
version = "1.0.0"
description = "YouTube Thumbnail A/B Testing Automation Tool"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "click>=8.1.7",
    "rich>=13.7.0",
    "httpx>=0.25.2",
    "apscheduler>=3.10.4",
    "sendgrid>=6.11.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.0.0",
]

[project.scripts]
yt-abtest = "app.main:cli"

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
